requirements.txt
boto3
botocore
simplejson
numpy
opencv-python
google-cloud